+++ src/structs/Transitions.jl
@@ -5,8 +5,8 @@
 =#
 
 struct Translation <: AbstractTransition
-    from::Union{Object,Point}
-    to::Union{Object,Point}
+    from::Union{Object,Luxor.AbstractPoint}
+    to::Union{Object,Luxor.AbstractPoint}
 end
 
 """
@@ -29,12 +29,15 @@ act!(obj, Action(1:50, anim_translate(10, 10)))
 """
 anim_translate(x::Real, y::Real) = anim_translate(Point(x, y))
 anim_translate(tp::Point) = Translation(O, tp)
-anim_translate(fp::Union{Object,Point}, tp::Union{Object,Point}) = Translation(fp, tp)
+anim_translate(
+    fp::Union{Object,Luxor.AbstractPoint},
+    tp::Union{Object,Luxor.AbstractPoint},
+) = Translation(fp, tp)
 
 struct Rotation{T<:Real} <: AbstractTransition
     from::T
     to::T
-    center::Union{Nothing,Point,AbstractObject}
+    center::Union{Nothing,Point,AbstractObject,DelayedPosition}
 end
 
 Rotation(from::Real, to::Real, center) = Rotation(promote(from, to)..., center)